+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/open-telemetry/opentelemetry-python/compare/v1.3.0-0.22b0...HEAD)
 
+### Added
+- Dropped attributes/events/links count available exposed on ReadableSpans.
+  ([#1893](https://github.com/open-telemetry/opentelemetry-python/pull/1893))
+- Added dropped count to otlp, jaeger and zipkin exporters.
+  ([#1893](https://github.com/open-telemetry/opentelemetry-python/pull/1893))
+
 ### Changed
 - Updated `opentelemetry-opencensus-exporter` to use `service_name` of spans instead of resource
   ([#1897](https://github.com/open-telemetry/opentelemetry-python/pull/1897))